// ui/src/generated/models/workItemType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Lintel
 * OpenAPI spec version: 0.1.0
 */

export type WorkItemType = typeof WorkItemType[keyof typeof WorkItemType];


export const WorkItemType = {
  feature: 'feature',
  bug: 'bug',
  refactor: 'refactor',
  task: 'task',
} as const;