// ui/src/generated/models/workItemStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Lintel
 * OpenAPI spec version: 0.1.0
 */

export type WorkItemStatus = typeof WorkItemStatus[keyof typeof WorkItemStatus];


export const WorkItemStatus = {
  open: 'open',
  in_progress: 'in_progress',
  in_review: 'in_review',
  approved: 'approved',
  merged: 'merged',
  closed: 'closed',
  failed: 'failed',
} as const;